page: Move "Package Format", pref -> page
- PreferencesDialog should only hold Add Water settings and not
    theme specific options
 - Fix freeze on Package format change. Now only refreshes profile
    combobox instead of whole GUI.
 - Apply action and profile combobox are disabled when selecting a
    package format that isn't installed.
 - Rename signal "refresh-gui" -> "package-changed"
 - Move profile combobox creation logic into its own method.
 - Remove _profile_cursor_reset() because it had weird behavior
    when combined with changing package formats. May reimplement it
    later if I can track all profiles across packages.
 - Move option factory into its own module `gui/option_factory.py`.
