+++ data/addwater.gresource.xml.in
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!-- TODO improve build to not use the prefix here but instead pass as a property at instantiation -->
 <gresources>
 	<gresource prefix="@PREFIX@">
 		<file preprocess="xml-stripblanks">gtk/window.ui</file>
@@ -6,7 +7,6 @@
 		<file preprocess="xml-stripblanks">gtk/preferences.ui</file>
 		<file preprocess="xml-stripblanks">gtk/addwater-page.ui</file>
 
-		<!-- TODO is there a benefit to making this app_id specific? Had trouble getting it to work-->
 		<file preprocess="xml-stripblanks" alias="dev.qwery.AddWater.metainfo">dev.qwery.AddWater.metainfo.xml.in</file>
 	</gresource>
