+++ data/gtk/preferences.ui
@@ -17,25 +17,6 @@
 				</child>
 				<child>
 					<object class="AdwPreferencesGroup">
-						<property name="title">Firefox</property>
-						<child>
-							<object class="AdwComboRow" id="firefox_package_combobox">
-								<property name="title" translatable="yes" comments="Translators: This refers to how the user installed Firefox (package manager, Flatpak, Snap, etc.)">Package Format</property>
-								<property name="tooltip-text"  translatable="yes">Profile data not found at selected location</property>
-								<property name="has-tooltip">false</property>
-								<property name="model">
-									<object class="GtkStringList" id="firefox_package_combobox_list">
-										<items>
-											<item translatable="yes">Automatically Discover</item>
-										</items>
-									</object>
-								</property>
-							</object>
-						</child>
-					</object>
-				</child>
-				<child>
-					<object class="AdwPreferencesGroup">
 						<property name="title" translatable="yes" comments="Translators: This refers to options that could be unsafe.">Danger Zone</property>
 						<property name="description" translatable="yes">Remove the theme from all known profiles and reset Add Water to its default state. Any custom changes made to your user.js file will remain after a reset.</property>
 						<child>
